+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Improved performance on `find_first` for small string arrays (ArrayString). This will improve the table name lookup
   performance.
 * Upgrade pegtl to 2.6.1. Several issues fixed.
+* Introduced Durability::Unsafe, which disables sync'ing to disk.
 
 ----------------------------------------------